Windows: Refactor the location of os_distinguisher
